refactor: delete target
