+++ src/components/pages/my-targets/DeleteTarget/DeleteTargetModal.styles.ts
