+++ src/components/pages/my-targets/DeleteTarget/DeleteTargetModal.tsx
