+++ src/pages/my-targets/index.tsx
@@ -1,9 +1,11 @@
+import { GetServerSideProps, GetServerSidePropsContext } from "next";
 import { getSession } from "next-auth/react";
 import { useEffect, useState } from "react";
 import PageTitle from "../../components/common/PageTitle/PageTitle";
 import Card from "../../components/pages/dashboard/Card/Card";
 import SafeProgress from "../../components/pages/dashboard/SafeProgress/SafeProgress";
 import CardNewTarget from "../../components/pages/my-targets/CardNewTarget/CardNewTarget";
+import DeleteTarget from "../../components/pages/my-targets/DeleteTarget/DeleteTarget";
 import {
   CardsContainer,
   Container,
@@ -11,35 +13,10 @@ import {
 import { getTargets as getTargetsService } from "../../services/getTargets";
 import { IGetTarget } from "../../types/IGetTarget";
 import { isValidToken } from "../../utils/isValidToken";
-import { GetServerSideProps, GetServerSidePropsContext } from "next";
-import { IError } from "../../types/IError";
-import Button from "../../components/common/Button/Button";
-import { isMobile } from "../../utils/isMobile";
-import { deleteTarget as deleteTargetService } from "../../services/deleteTarget";
-import Dialog from "../../components/common/Dialog/Dialog";
-import DeleteTargetModal from "../../components/pages/my-targets/DeleteTargetModal/DeleteTargetModal";
-import Feedback from "../../components/common/Feedback/Feedback";
-import { IAlertProps } from "../../types/IAlertProps";
-import Router from "next/router";
-
-const defaultError = {
-  hasError: false,
-  message: "",
-};
-
-const defaultAlert: IAlertProps = {
-  severity: "success",
-  message: "Aporte excluído com sucesso!",
-};
 
 const MyTargets = (): JSX.Element => {
-  const mobile = isMobile();
   const [targets, setTargets] = useState<IGetTarget[] | null>(null);
   const [deleteModalOpened, setDeleteModalOpened] = useState<boolean>(false);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [buttonDisabled, setButtonDisabled] = useState<boolean>(false);
-  const [feedbackOpened, setFeedbackOpened] = useState<boolean>(false);
-  const [error, setError] = useState<IError>(defaultError);
   const [deleteTargetData, setDeleteTargetData] = useState<IGetTarget | null>(
     null
   );
@@ -65,48 +42,6 @@ const MyTargets = (): JSX.Element => {
     setDeleteTargetData(target);
   };
 
-  const handleFeedbackClose = () => {
-    setDeleteModalOpened(false);
-    Router.reload();
-  };
-
-  const deleteTarget = async (data: { target_id: string }): Promise<void> => {
-    setIsLoading(true);
-    setButtonDisabled(true);
-
-    const session = await getSession();
-    const response = await deleteTargetService(data, session?.user.jwt);
-
-    if (response && response.success) {
-      setFeedbackOpened(true);
-    } else {
-      setError({
-        hasError: true,
-        message: response.message,
-      });
-    }
-  };
-
-  const DialogButtons = (
-    <>
-      <Button
-        type="button"
-        size={mobile ? "medium" : "small"}
-        text="Não"
-        outlined
-        onClick={(): void => setDeleteModalOpened(false)}
-      />
-      <Button
-        type="submit"
-        form="deleteTargetForm"
-        size={mobile ? "medium" : "small"}
-        text="Confirmar"
-        loading={isLoading}
-        disabled={buttonDisabled}
-      />
-    </>
-  );
-
   return (
     <Container>
       <PageTitle>Meus objetivos</PageTitle>
@@ -130,28 +65,12 @@ const MyTargets = (): JSX.Element => {
             )
           )}
       </CardsContainer>
-      {deleteTargetData && (
-        <>
-          <Dialog
-            isOpen={deleteModalOpened}
-            title="Deseja realmente remover este objetivo?"
-            handleClose={(): void => setDeleteModalOpened(false)}
-            buttons={DialogButtons}
-          >
-            <DeleteTargetModal
-              data={deleteTargetData}
-              onSubmit={deleteTarget}
-              error={error}
-            />
-          </Dialog>
 
-          <Feedback
-            feedbackOpened={feedbackOpened}
-            alertProps={defaultAlert}
-            handleClose={handleFeedbackClose}
-          />
-        </>
-      )}
+      <DeleteTarget
+        deleteTargetData={deleteTargetData}
+        deleteModalOpened={deleteModalOpened}
+        setDeleteModalOpened={setDeleteModalOpened}
+      />
     </Container>
   );
 };
